// web/lib/api/generated/model/rule.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationChannelEnum } from './notificationChannelEnum';
import type { NotificationRecipientEnum } from './notificationRecipientEnum';

export interface Rule {
  recipient: NotificationRecipientEnum;
  /** @maxLength 60 */
  permission?: string;
  /** @maxItems 4 */
  channels: NotificationChannelEnum[];
  enabled?: boolean;
  compulsory?: boolean;
  readonly is_default: boolean;
}